Added logout to main and balance windows. Also added card status check before calling pincode and updated get_card_status to use long long as card_num type.

+++ QT_frontend/DLLRestAPI/dll_rest_api.cpp
@@ -30,7 +30,7 @@ dll_rest_api::~dll_rest_api() {
 void dll_rest_api::login(QByteArray hash, long long card_num) { p_engine->login(hash, card_num); }
 void dll_rest_api::get_customer_info(int customer_id) { p_engine->customer_info(customer_id); }
 void dll_rest_api::get_account_info(int account_id) { p_engine->account_info(account_id); }
-void dll_rest_api::get_card_status(int card_num) { p_engine->card_status(card_num); }
+void dll_rest_api::get_card_status(long long card_num) { p_engine->card_status(card_num); }
 void dll_rest_api::get_transactions(int account_id, int index) {
     p_engine->transactions(account_id, index);
 }
